+++ nes-emu-ios/nes-emu-ios/Model/NES/Mappers/Mapper_NTDEC2722.swift
@@ -50,7 +50,7 @@ struct Mapper_NTDEC2722: MapperProtocol
     {
         if let safeState = aState
         {
-            self.mirroringMode = MirroringMode.init(rawValue: safeState.mirroringMode) ?? aCartridge.header.mirroringMode
+            self.mirroringMode = MirroringMode.init(rawValue: Int(safeState.mirroringMode)) ?? aCartridge.header.mirroringMode
             self.prgBank = safeState.ints[safe: 0] ?? 0
             self.cycles = safeState.ints[safe: 1] ?? 0
             
@@ -84,11 +84,11 @@ struct Mapper_NTDEC2722: MapperProtocol
     {
         get
         {
-            MapperState(mirroringMode: self.mirroringMode.rawValue, ints: [self.prgBank, self.cycles], bools: [], uint8s: [], chr: self.chr)
+            MapperState(mirroringMode: UInt8(self.mirroringMode.rawValue), ints: [self.prgBank, self.cycles], bools: [], uint8s: [], chr: self.chr)
         }
         set
         {
-            self.mirroringMode = MirroringMode.init(rawValue: newValue.mirroringMode) ?? self.mirroringMode
+            self.mirroringMode = MirroringMode.init(rawValue: Int(newValue.mirroringMode)) ?? self.mirroringMode
             self.prgBank = newValue.ints[safe: 0] ?? 0
             self.cycles = newValue.ints[safe: 1] ?? 0
             self.chr = newValue.chr
